session_number applies to ophys session and experiments
