+++ allensdk/brain_observatory/behavior/behavior_project_cache/behavior_sessions_cache.py
@@ -9,7 +9,7 @@ from allensdk.brain_observatory.behavior.behavior_project_cache.cache_table impo
 class BehaviorSessionsCacheTable(CacheTable):
     def __init__(self, df: pd.DataFrame,
                  suppress: Optional[List[str]] = None):
-        super().__init__(df=df, suppress=suppress)
+        super().__init__(df=df, suppress=suppress, behavior_only=True)
 
     def postprocess_additional(self):
         pass
